update pixel field format

+++ contracts/Controller.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.8.9;
 // import "hardhat/console.sol";
 
 contract Controller {
-    uint[256] public pixels;
+    uint8[256] public pixels;
 
     event PixelChanged(uint x, uint y, uint color);
 
@@ -19,11 +19,11 @@ contract Controller {
         return pixels[x + y * 16];
     }
 
-    function getAllPixels() public view returns (uint[256] memory) {
+    function getAllPixels() public view returns (uint8[256] memory) {
         return pixels;
     }
 
-    function setPixel(uint x, uint y, uint color) public {
+    function setPixel(uint x, uint y, uint8 color) public {
         pixels[x + y * 16] = color;
         emit PixelChanged(x, y, color);
     }
